organize exchange id methods in class
